refactor a transformer
